moved cloud_client clb functions to clb.py
and changed all the callers to point to new place
