+++ otter/cloud_client/__init__.py
@@ -2,7 +2,6 @@
 import json
 import re
 from functools import partial, wraps
-from operator import itemgetter
 from urlparse import parse_qs, urlparse
 
 import attr
@@ -15,15 +14,12 @@ from effect import (
     TypeDispatcher,
     catch,
     perform,
-    raise_,
     sync_performer)
-from effect.do import do, do_return
 
 import six
 
 from toolz.dicttoolz import get_in
 from toolz.functoolz import identity
-from toolz.itertoolz import concat
 
 from twisted.internet.task import deferLater
 
@@ -31,10 +27,9 @@ from txeffect import deferred_performer, perform as twisted_perform
 
 from otter.auth import Authenticate, InvalidateToken, public_endpoint_url
 from otter.constants import ServiceType
-from otter.indexer import atom
 from otter.log.intents import msg as msg_effect
 from otter.util.config import config_value
-from otter.util.http import APIError, append_segments, try_json_with_keys
+from otter.util.http import APIError, append_segments
 from otter.util.http import headers as otter_headers
 from otter.util.pure_http import (
     add_bind_root,
@@ -372,37 +367,6 @@ def log_success_response(msg_type, response_body_filter, log_as_json=True,
     return _log_it
 
 
-# ----- CLB requests and error parsing -----
-def _regex(pattern):
-    """
-    Compile a case-insensitive pattern.
-    """
-    return re.compile(pattern, re.I)
-
-
-_CLB_IMMUTABLE_PATTERN = _regex(
-    "Load\s*Balancer '\d+' has a status of '[^']+' and is considered "
-    "immutable")
-_CLB_NOT_ACTIVE_PATTERN = _regex("Load\s*Balancer is not ACTIVE")
-_CLB_DELETED_PATTERN = _regex(
-    "(Load\s*Balancer '\d+' has a status of 'PENDING_DELETE' and is|"
-    "The load balancer is deleted and) considered immutable")
-_CLB_MARKED_DELETED_PATTERN = _regex(
-    "The load\s*balancer is marked as deleted")
-_CLB_NO_SUCH_NODE_PATTERN = _regex(
-    "Node with id #\d+ not found for load\s*balancer #\d+$")
-_CLB_NO_SUCH_LB_PATTERN = _regex(
-    "Load\s*balancer not found")
-_CLB_DUPLICATE_NODES_PATTERN = _regex(
-    "Duplicate nodes detected. One or more nodes already configured "
-    "on load\s*balancer")
-_CLB_NODE_LIMIT_PATTERN = _regex(
-    "Nodes must not exceed (\d+) per load\s*balancer")
-_CLB_NODE_REMOVED_PATTERN = _regex(
-    "Node ids ((?:\d+,)*(?:\d+)) are not a part of your load\s*balancer")
-_CLB_OVER_LIMIT_PATTERN = _regex("OverLimit Retry\.{3}")
-
-
 @attr.s(these={"message": attr.ib()}, init=False)
 class ExceptionWithMessage(Exception):
     """
@@ -430,96 +394,7 @@ class ExceptionWithMessage(Exception):
         self.message = message
 
 
-@attributes([Attribute('lb_id', instance_of=six.text_type)])
-class CLBImmutableError(Exception):
-    """
-    Error to be raised when the CLB is in some status that causes is to be
-    temporarily immutable.
-
-    This exception is _not_ used when the status is PENDING_DELETE. See
-    :obj:`CLBDeletedError`.
-    """
-
-
-@attributes([Attribute('lb_id', instance_of=six.text_type)])
-class CLBNotFoundError(Exception):
-    """A CLB doesn't exist. Superclass of other, more specific exceptions."""
-
-
-class CLBDeletedError(CLBNotFoundError):
-    """
-    Error to be raised when the CLB has been deleted or is being deleted.
-    This is distinct from it not existing.
-    """
-
-
-class NoSuchCLBError(CLBNotFoundError):
-    """
-    Error to be raised when the CLB never existed in the first place (or it
-    has been deleted so long that there is no longer a record of it).
-    """
-
-
-@attributes([Attribute('lb_id', instance_of=six.text_type),
-             Attribute('node_id', instance_of=six.text_type)])
-class NoSuchCLBNodeError(Exception):
-    """
-    Error to be raised when attempting to modify a CLB node that no longer
-    exists.
-    """
-
-
-@attributes([Attribute('lb_id', instance_of=six.text_type)])
-class CLBNotActiveError(Exception):
-    """
-    Error to be raised when a CLB is not ACTIVE (and we have no more
-    information about what its actual state is).
-    """
-
-
-@attributes([Attribute('lb_id', instance_of=six.text_type)])
-class CLBRateLimitError(Exception):
-    """
-    Error to be raised when CLB returns 413 (rate limiting).
-    """
-
-
-@attributes([Attribute('lb_id', instance_of=six.text_type)])
-class CLBDuplicateNodesError(Exception):
-    """
-    Error to be raised only when adding one or more nodes to a CLB whose
-    address and port are mapped on the CLB.
-    """
-
-
-@attributes([Attribute('lb_id', instance_of=six.text_type),
-             Attribute("node_limit", instance_of=int)])
-class CLBNodeLimitError(Exception):
-    """
-    Error to be raised only when adding one or more nodes to a CLB: adding
-    that number of nodes would exceed the maximum number of nodes allowed on
-    the CLB.
-    """
-
-
-@attr.s
-class CLBPartialNodesRemoved(Exception):
-    """
-    Exception raised when only some of the nodes are removed.
-
-    :ivar lb_id: CLB ID
-    :type: :obj:`six.text_type`
-    :ivar list not_removed_node_ids: List of node_ids not removed where each
-        node_id is :obj:`six.text_type`
-    :ivar list removed_node_ids: List of node_ids removed where each node_id
-        is :obj:`six.text_type`
-    """
-    lb_id = attr.ib(validator=attr.validators.instance_of(six.text_type))
-    not_removed_node_ids = attr.ib(validator=attr.validators.instance_of(list))
-    removed_node_ids = attr.ib(validator=attr.validators.instance_of(list))
-
-
-def _match_errors(code_keys_exc_mapping, status_code, response_dict):
+def match_errors(code_keys_exc_mapping, status_code, response_dict):
     """
     Take a list of tuples of:
     (status code, json keys, regex pattern (optional), exception callable),
@@ -534,7 +409,7 @@ def _match_errors(code_keys_exc_mapping, status_code, response_dict):
                 raise make_exc(message)
 
 
-def _only_json_api_errors(f):
+def only_json_api_errors(f):
     """
     Helper function so that we only catch APIErrors with bodies that can be
     parsed into JSON.
@@ -560,313 +435,17 @@ def _only_json_api_errors(f):
     return catch(APIError, try_parsing)
 
 
-def add_clb_nodes(lb_id, nodes):
+def regex(pattern):
     """
-    Generate effect to add one or more nodes to a load balancer.
-
-    Note: This is not correctly documented in the load balancer documentation -
-    it is documented as "Add Node" (singular), but the examples show multiple
-    nodes being added.
-
-    :param str lb_id: The load balancer ID to add the nodes to
-    :param list nodes: A list of node dictionaries that each look like::
-
-        {
-            "address": "valid ip address",
-            "port": 80,
-            "condition": "ENABLED",
-            "weight": 1,
-            "type": "PRIMARY"
-        }
-
-        (weight and type are optional)
-
-    :return: :class:`ServiceRequest` effect
-
-    :raises: :class:`CLBImmutableError`, :class:`CLBDeletedError`,
-        :class:`NoSuchCLBError`, :class:`CLBDuplicateNodesError`,
-        :class:`APIError`
+    Compile a case-insensitive pattern.
     """
-    eff = service_request(
-        ServiceType.CLOUD_LOAD_BALANCERS,
-        'POST',
-        append_segments('loadbalancers', lb_id, 'nodes'),
-        data={'nodes': nodes},
-        success_pred=has_code(202))
-
-    @_only_json_api_errors
-    def _parse_known_errors(code, json_body):
-        mappings = _expand_clb_matches(
-            [(422, _CLB_DUPLICATE_NODES_PATTERN, CLBDuplicateNodesError)],
-            lb_id)
-        _match_errors(mappings, code, json_body)
-        _process_clb_api_error(code, json_body, lb_id)
-        process_nodelimit_error(code, json_body, lb_id)
-
-    return eff.on(error=_parse_known_errors).on(
-        log_success_response('request-add-clb-nodes', identity))
-
-
-def process_nodelimit_error(code, json_body, lb_id):
-    """
-    Parse error that causes CLBNodeLimitError along with limit and raise it
-    """
-    if code != 413:
-        return
-    match = _CLB_NODE_LIMIT_PATTERN.match(json_body.get("message", ""))
-    if match is not None:
-        limit = int(match.group(1))
-        raise CLBNodeLimitError(lb_id=six.text_type(lb_id), node_limit=limit)
-
-
-def change_clb_node(lb_id, node_id, condition, weight, _type="PRIMARY"):
-    """
-    Generate effect to change a node on a load balancer.
-
-    :param str lb_id: The load balancer ID to add the nodes to
-    :param str node_id: The node id to change.
-    :param str condition: The condition to change to: one of "ENABLED",
-        "DRAINING", or "DISABLED"
-    :param int weight: The weight to change to.
-    :param str _type: The type to change the CLB node to.
-
-    :return: :class:`ServiceRequest` effect
-
-    :raises: :class:`CLBImmutableError`, :class:`CLBDeletedError`,
-        :class:`NoSuchCLBError`, :class:`NoSuchCLBNodeError`, :class:`APIError`
-    """
-    eff = service_request(
-        ServiceType.CLOUD_LOAD_BALANCERS,
-        'PUT',
-        append_segments('loadbalancers', lb_id, 'nodes', node_id),
-        data={'node': {
-            'condition': condition, 'weight': weight, 'type': _type}},
-        success_pred=has_code(202))
-
-    @_only_json_api_errors
-    def _parse_known_errors(code, json_body):
-        _process_clb_api_error(code, json_body, lb_id)
-        _match_errors(
-            _expand_clb_matches(
-                [(404, _CLB_NO_SUCH_NODE_PATTERN, NoSuchCLBNodeError)],
-                lb_id=lb_id, node_id=node_id),
-            code,
-            json_body)
-
-    return eff.on(error=_parse_known_errors)
-    # CLB 202 response here has no body, so no response logging needed
-
-
-# Number of nodes that can be deleted in `DELETE ../nodes` call as per
-# https://developer.rackspace.com/docs/cloud-load-balancers/v1/api-reference/nodes/#bulk-delete-nodes
-CLB_BATCH_DELETE_LIMIT = 10
-
-
-def remove_clb_nodes(lb_id, node_ids):
-    """
-    Remove multiple nodes from a load balancer.
-
-    :param str lb_id: A load balancer ID.
-    :param node_ids: iterable of node IDs.
-    :return: Effect of None.
-
-    Succeeds on 202.
-
-    This function will handle the case where *some* of the nodes are valid and
-    some aren't, by retrying deleting only the valid ones.
-    """
-    node_ids = list(node_ids)
-    partial = None
-    if len(node_ids) > CLB_BATCH_DELETE_LIMIT:
-        not_removing = node_ids[CLB_BATCH_DELETE_LIMIT:]
-        node_ids = node_ids[:CLB_BATCH_DELETE_LIMIT]
-        partial = CLBPartialNodesRemoved(six.text_type(lb_id),
-                                         map(six.text_type, not_removing),
-                                         map(six.text_type, node_ids))
-    eff = service_request(
-        ServiceType.CLOUD_LOAD_BALANCERS,
-        'DELETE',
-        append_segments('loadbalancers', lb_id, 'nodes'),
-        params={'id': map(str, node_ids)},
-        success_pred=has_code(202))
-
-    def check_invalid_nodes(exc_info):
-        code = exc_info[1].code
-        body = exc_info[1].body
-        if code == 400:
-            message = try_json_with_keys(
-                body, ["validationErrors", "messages", 0])
-            if message is not None:
-                match = _CLB_NODE_REMOVED_PATTERN.match(message)
-                if match:
-                    removed = concat([group.split(',')
-                                      for group in match.groups()])
-                    return remove_clb_nodes(lb_id,
-                                            set(node_ids) - set(removed))
-        six.reraise(*exc_info)
-
-    return eff.on(
-        error=catch(APIError, check_invalid_nodes)
-    ).on(
-        error=_only_json_api_errors(
-            lambda c, b: _process_clb_api_error(c, b, lb_id))
-    ).on(success=lambda _: None if partial is None else raise_(partial))
-    # CLB 202 responses here has no body, so no response logging needed.
-
-
-def get_clb_nodes(lb_id):
-    """
-    Fetch the nodes of the given load balancer. Returns list of node JSON.
-    """
-    return service_request(
-        ServiceType.CLOUD_LOAD_BALANCERS,
-        'GET',
-        append_segments('loadbalancers', str(lb_id), 'nodes'),
-    ).on(
-        error=_only_json_api_errors(
-            lambda c, b: _process_clb_api_error(c, b, lb_id))
-    ).on(
-        log_success_response('request-list-clb-nodes', identity)
-    ).on(
-        success=lambda (response, body): body['nodes'])
-
-
-def get_clbs():
-    """Fetch all LBs for a tenant. Returns list of loadbalancer JSON."""
-    return service_request(
-        ServiceType.CLOUD_LOAD_BALANCERS, 'GET', 'loadbalancers',
-    ).on(
-        log_success_response('request-list-clbs', identity)
-    ).on(
-        success=lambda (response, body): body['loadBalancers'])
-
-
-def _node_feed_page(lb_id, node_id, params):
-    """
-    Return page of CLB node feed
-
-    :param int lb_id: Cloud Load balancer ID
-    :param int node_id: Node ID of in loadbalancer node
-    :param dict params: Request query parameters
-
-    :returns: Unparsed response body as string
-    :rtype: `str`
-    """
-    return service_request(
-        ServiceType.CLOUD_LOAD_BALANCERS, 'GET',
-        append_segments('loadbalancers', str(lb_id), 'nodes',
-                        '{}.atom'.format(node_id)),
-        params=params,
-        json_response=False
-    ).on(
-        error=_only_json_api_errors(
-            lambda c, b: _process_clb_api_error(c, b, lb_id))
-    ).on(
-        log_success_response('request-get-clb-node-feed', identity)
-    ).on(itemgetter(1))
-
-
-@do
-def get_clb_node_feed(lb_id, node_id):
-    """
-    Get the atom feed associated with a CLB node.
-
-    :param int lb_id: Cloud Load balancer ID
-    :param int node_id: Node ID of in loadbalancer node
-
-    :returns: Effect of ``list`` of atom entry :class:`Element`
-    :rtype: ``Effect``
-    """
-    all_entries = []
-    params = {}
-    while True:
-        feed_str = yield _node_feed_page(lb_id, node_id, params)
-        feed = atom.parse(feed_str)
-        entries = atom.entries(feed)
-        if entries == []:
-            break
-        all_entries.extend(entries)
-        next_link = atom.next_link(feed)
-        if not next_link:
-            break
-        params = parse_qs(urlparse(next_link).query)
-    yield do_return(all_entries)
-
-
-def get_clb_health_monitor(lb_id):
-    """
-    Return CLB health monitor setting
-
-    :param int lb_id: Loadbalancer ID
-
-    :return: ``Effect`` of ``dict`` representing health monitor config
-    """
-    return service_request(
-        ServiceType.CLOUD_LOAD_BALANCERS,
-        'GET',
-        append_segments('loadbalancers', str(lb_id), 'healthmonitor')
-    ).on(
-        error=_only_json_api_errors(
-            lambda c, b: _process_clb_api_error(c, b, lb_id))
-    ).on(
-        log_success_response('request-get-clb-healthmon', identity)
-    ).on(
-        success=lambda (response, body): body["healthMonitor"])
-
-
-def _expand_clb_matches(matches_tuples, lb_id, node_id=None):
-    """
-    All CLB messages have only the keys ("message",), and the exception tpye
-    takes a load balancer ID and maybe a node ID.  So expand a tuple that looks
-    like:
-
-    (code, pattern, exc_type)
-
-    to
-
-    (code, ("message",), pattern, partial(exc_type, lb_id=lb_id))
-
-    and maybe the partial will include the node ID too if it's provided.
-    """
-    params = {"lb_id": six.text_type(lb_id)}
-    if node_id is not None:
-        params["node_id"] = six.text_type(node_id)
-
-    return [(m[0], ("message",), m[1], partial(m[2], **params))
-            for m in matches_tuples]
-
-
-def _process_clb_api_error(api_error_code, json_body, lb_id):
-    """
-    Attempt to parse generic CLB API error messages, and raise recognized
-    exceptions in their place.
-
-    :param int api_error_code: The status code from the HTTP request
-    :param dict json_body: The error message, parsed as a JSON dict.
-    :param string lb_id: The load balancer ID
-
-    :raises: :class:`CLBImmutableError`, :class:`CLBDeletedError`,
-        :class:`NoSuchCLBError`, :class:`APIError` by itself
-    """
-    mappings = (
-        # overLimit is different than the other CLB messages because it's
-        # produced by repose
-        [(413, ("overLimit", "message"), _CLB_OVER_LIMIT_PATTERN,
-          partial(CLBRateLimitError, lb_id=six.text_type(lb_id)))] +
-        _expand_clb_matches(
-            [(422, _CLB_DELETED_PATTERN, CLBDeletedError),
-             (410, _CLB_MARKED_DELETED_PATTERN, CLBDeletedError),
-             (422, _CLB_IMMUTABLE_PATTERN, CLBImmutableError),
-             (422, _CLB_NOT_ACTIVE_PATTERN, CLBNotActiveError),
-             (404, _CLB_NO_SUCH_LB_PATTERN, NoSuchCLBError)],
-            lb_id))
-    return _match_errors(mappings, api_error_code, json_body)
+    return re.compile(pattern, re.I)
 
 
 # ----- Nova requests and error parsing -----
 
 def _forbidden_plaintext(message):
-    return _regex(
+    return regex(
         "403 Forbidden\s+Access was denied to this resource\.\s+({0})\s*$"
         .format(message))
 
@@ -877,7 +456,7 @@ _NOVA_403_PUBLIC_SERVICENET_BOTH_REQUIRED = _forbidden_plaintext(
     "11111111-1111-1111-1111-111111111111\) required but missing")
 _NOVA_403_RACKCONNECT_NETWORK_REQUIRED = _forbidden_plaintext(
     "Exactly 1 isolated network\(s\) must be attached")
-_NOVA_403_QUOTA = _regex(
+_NOVA_403_QUOTA = regex(
     "Quota exceeded for (\S+): Requested \d+, but already used \d+ of \d+ "
     "(\S+)")
 
@@ -926,7 +505,7 @@ class CreateServerOverQuoteError(ExceptionWithMessage):
     """
 
 
-_MAX_METADATA_PATTERN = _regex('Maximum number of metadata items')
+_MAX_METADATA_PATTERN = regex('Maximum number of metadata items')
 
 
 def set_nova_metadata_item(server_id, key, value):
@@ -952,7 +531,7 @@ def set_nova_metadata_item(server_id, key, value):
         reauth_codes=(401,),
         success_pred=has_code(200))
 
-    @_only_json_api_errors
+    @only_json_api_errors
     def _parse_known_errors(code, json_body):
         other_errors = [
             (404, ('itemNotFound', 'message'), None,
@@ -961,7 +540,7 @@ def set_nova_metadata_item(server_id, key, value):
              partial(ServerMetadataOverLimitError,
                      server_id=six.text_type(server_id))),
         ]
-        _match_errors(_nova_standard_errors + other_errors, code, json_body)
+        match_errors(_nova_standard_errors + other_errors, code, json_body)
 
     return eff.on(error=_parse_known_errors).on(
         log_success_response('request-set-metadata-item', identity))
@@ -985,13 +564,13 @@ def get_server_details(server_id):
         append_segments('servers', server_id),
         success_pred=has_code(200))
 
-    @_only_json_api_errors
+    @only_json_api_errors
     def _parse_known_errors(code, json_body):
         other_errors = [
             (404, ('itemNotFound', 'message'), None,
              partial(NoSuchServerError, server_id=six.text_type(server_id))),
         ]
-        _match_errors(_nova_standard_errors + other_errors, code, json_body)
+        match_errors(_nova_standard_errors + other_errors, code, json_body)
 
     return eff.on(error=_parse_known_errors).on(
         log_success_response('request-one-server-details', identity))
@@ -1020,7 +599,7 @@ def create_server(server_args):
         success_pred=has_code(202),
         reauth_codes=(401,))
 
-    @_only_json_api_errors
+    @only_json_api_errors
     def _parse_known_json_errors(code, json_body):
         other_errors = [
             (400, ('badRequest', 'message'), None,
@@ -1028,7 +607,7 @@ def create_server(server_args):
             (403, ('forbidden', 'message'), _NOVA_403_QUOTA,
              CreateServerOverQuoteError)
         ]
-        _match_errors(_nova_standard_errors + other_errors, code, json_body)
+        match_errors(_nova_standard_errors + other_errors, code, json_body)
 
     def _parse_known_string_errors(api_error_exc_info):
         api_error = api_error_exc_info[1]
@@ -1068,9 +647,9 @@ def list_servers_details_page(parameters=None):
     :raise: :class:`NovaRateLimitError`, :class:`NovaComputeFaultError`,
         :class:`APIError`
     """
-    @_only_json_api_errors
+    @only_json_api_errors
     def _parse_known_errors(code, json_body):
-        _match_errors(_nova_standard_errors, code, json_body)
+        match_errors(_nova_standard_errors, code, json_body)
 
     return (
         service_request(
