+++ otter/convergence/errors.py
@@ -8,10 +8,12 @@ from toolz.functoolz import identity
 
 from otter.auth import NoSuchEndpoint
 from otter.cloud_client import (
+    CreateServerConfigurationError,
+    CreateServerOverQuoteError
+)
+from otter.cloud_client.clb import (
     CLBDeletedError,
     CLBNodeLimitError,
-    CreateServerConfigurationError,
-    CreateServerOverQuoteError,
     NoSuchCLBError,
     NoSuchCLBNodeError
 )
